pyproject.toml
[project]
name = "vla0"
version = "0.1.0"
description = "Minimal VLA-0 implementation for LIBERO"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.38.0",
    "qwen-vl-utils",
    "accelerate",
    "h5py",
    "pillow",
    "wandb",
    "tqdm",
    "numpy",
    "bitsandbytes",
    "peft>=0.7.0",
]

[project.optional-dependencies]
eval = [
    "libero @ git+https://github.com/Lifelong-Robot-Learning/LIBERO.git",
]

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
